+++ odoo/songs/upgrade/10_22_2_update.py
@@ -23,6 +23,16 @@ def _recompute_sale_order_line(ctx):
 
 
 @anthem.log
+def update_esb_cron(ctx):
+    """  Set up esb cron to use queue_job_cron."""
+    sql = ("UPDATE ir_cron SET run_as_queue_job=True, channel_id={}"
+           "    WHERE model='connector_esb'".format(
+               ctx.env.ref('queue_job_cron.channel_root_ir_cron').id))
+    ctx.env.cr.execute(sql)
+
+
+@anthem.log
 def main(ctx):
     _recompute_sale_order_line(ctx)
     uninstall_module_specific_translations(ctx)
+    update_esb_cron(ctx)